+++ src/index.ts
@@ -2,67 +2,10 @@ import chalk from "chalk";
 import fs from "fs";
 
 import { getHelp, Namespace, parseArgs } from "./cli";
-import {
-    getLocalPath,
-    getWorkflowTitle,
-    readLocalWorkflows,
-    readRemoteWorkflows,
-    updateWorkflow
-} from "./manager";
-import { decapitalize } from "./utils";
-
-function processWorkflowUpdate(
-    name: string,
-    localContent: string | null,
-    remoteContent: string | null,
-    updateExisting: boolean
-): boolean {
-    const localPath = getLocalPath(name);
-    const workflowPrefix = `Workflow ${chalk.bold(localPath)}`;
-    if (!remoteContent) {
-        console.warn(
-            chalk.red(`✗  ${workflowPrefix} download failed, skipping`)
-        );
-        return false;
-    }
-    const workflowPurpose = chalk.bold(
-        decapitalize(getWorkflowTitle(remoteContent))
-    );
-    if (!localContent) {
-        updateWorkflow(name, remoteContent);
-        console.info(
-            chalk.green(
-                `✓  Workflow ${chalk.bold(
-                    localPath
-                )} added to ${workflowPurpose}`
-            )
-        );
-        return true;
-    }
-    if (localContent === remoteContent) {
-        console.log(
-            chalk.grey(
-                `✓  ${workflowPrefix} is up to date and ready to ${workflowPurpose}`
-            )
-        );
-        return false;
-    }
-    if (!updateExisting) {
-        console.info(
-            chalk.yellow(
-                `↻  ${workflowPrefix} can ${workflowPurpose} better, add ${chalk.bold(
-                    "-f"
-                )} CLI flag to update`
-            )
-        );
-        return false;
-    }
-    updateWorkflow(name, remoteContent);
-    console.info(
-        chalk.green(`✓  ${workflowPrefix} updated, cross-check changes`)
-    );
-    return true;
-}
+import { WORKFLOW_NAMES } from "./constants";
+import { readLocalWorkflows, readRemoteWorkflows } from "./manager";
+import { runCheckAll } from "./runCheck";
+import runUpdate from "./runUpdate";
 
 async function main(): Promise<void> {
     let args: Namespace;
@@ -97,6 +40,12 @@ async function main(): Promise<void> {
         );
         process.exit(1);
     }
+    if (args.check) {
+        const names = args.update.length ? args.update : WORKFLOW_NAMES;
+        const result = await runCheckAll(names, args.ref);
+        process.exit(result ? 0 : 1);
+    }
+
     if (!args.update.length) {
         console.log(getHelp());
         process.exit(0);
@@ -111,7 +60,7 @@ async function main(): Promise<void> {
     const localContents = new Map(await readLocalWorkflows(args.update));
     remoteContents.forEach(([name, remoteContent]) => {
         const localContent = localContents.get(name) || null;
-        processWorkflowUpdate(name, localContent, remoteContent, args.force);
+        runUpdate(name, localContent, remoteContent, args.force);
     });
 }