+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `[cli]` `--check` flag to check if workflows are ready for updates
+
 ### Fixed
 - Check if CLI command is started in a GitHub repository root
 - `[on_push_check]` `npm run` commands `lint` and `test` were not discovered properly from 3rd party scripts